Add middleware and refactor blog controller
Create middleware 'userExtract' that given a token
it add a user object to the request.

Add the middleware to the post and delete endpoints
in the blog controller.

Refactor blog controller, using the new middleware.

On deletion of a blog, also delete it from the user
object blog property and save it.